+++ src/jest.setup.ts
@@ -1,7 +1,7 @@
 import { connectDB, disconnectDB } from './config';
 import { startServer, stopServer } from './server';
 import { envConstants } from './constants';
-import { setupInitialRolesAndPermissions, setupInitialUserRole } from './script/initialize';
+import { setupInitialRolesAndPermissions, setupInitialUserRole } from './script';
 
 beforeAll(async () => {
   await connectDB();
