+++ src/middlewares/axiosHeader.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from 'express';
 import axios from 'axios';
-import { getServerURL } from '../utils/helper.util';
+import { getServerURL } from '@/utils';
 
 export const addAxiosHeadersMiddleware = (req: Request, _: Response, next: NextFunction) => {
   const serverURL = getServerURL(req);
