+++ src/middlewares/index.ts
@@ -2,3 +2,4 @@ export * from './auth';
 export * from './errorHandler';
 export * from './passport';
 export * from './checkPermission';
+export * from './axiosHeader';
