+++ src/services/index.ts
@@ -1,0 +1,2 @@
+export * from './notification.validation';
+export * from './notification.services';
